update recipe generate logic
