--- src/main/java/me/myogoo/extendedterminal/api/DataGen.java
@@ -1,4 +1,0 @@
-package me.myogoo.extendedterminal.api;
-
-public @interface DataGen {
-}
